+++ src/components/callStatement.cc
@@ -81,8 +81,8 @@ pair<
 	return make_pair(this->elements.begin(), this->elements.end());
 }
 
-unsigned int CallStatement::getElementCount() {
-	unsigned int count = 0;
+size_t CallStatement::getElementCount() {
+	size_t count = 0;
 	for(CallElement element: this->elements) {
 		if(element.component != nullptr) {
 			count++;
