+++ src/tssl/echo.h
@@ -6,7 +6,7 @@ using namespace std;
 
 namespace ts {
 	namespace sl {
-		void* echo(int argc, void** args) {
+		void* echo(size_t argc, void** args) {
 			if(argc >= 1) {
 				printf("%s\n", ((string*)args[0])->c_str());
 			}
