+++ src/tssl/scriptObject.cc
@@ -2,7 +2,7 @@
 
 using namespace ts;
 
-void* sl::ScriptObject__test(int argc, void** args) {
+void* sl::ScriptObject__test(size_t argc, void** args) {
 	if(argc >= 1) {
 		printf("script object says: %s\n", ((string*)args[1])->c_str());
 	}
